// frontend/src/lib/generated/api/models/DTOProfesor.ts
export interface DTOProfesor {
    /**
     * Unique professor ID
     * @type {number}
     * @memberof DTOProfesor
     */
    id?: number;
    /**
     * Unique username
     * @type {string}
     * @memberof DTOProfesor
     */
    username: string;
    /**
     * Professor's first name
     * @type {string}
     * @memberof DTOProfesor
     */
    firstName: string;
    /**
     * Professor's last name
     * @type {string}
     * @memberof DTOProfesor
     */
    lastName: string;
    /**
     * Professor's DNI
     * @type {string}
     * @memberof DTOProfesor
     */
    dni: string;
    /**
     * Professor's email
     * @type {string}
     * @memberof DTOProfesor
     */
    email: string;
    /**
     * Phone number
     * @type {string}
     * @memberof DTOProfesor
     */
    phoneNumber?: string;
    /**
     * User role
     * @type {string}
     * @memberof DTOProfesor
     */
    role?: DTOProfesorRoleEnum;
    /**
     * Indicates if the account is enabled
     * @type {boolean}
     * @memberof DTOProfesor
     */
    enabled?: boolean;
    /**
     * List of assigned class IDs
     * @type {Array<number>}
     * @memberof DTOProfesor
     */
    classIds?: Array<number>;
    /**
     * Profile creation date
     * @type {Date}
     * @memberof DTOProfesor
     */
    createdAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof DTOProfesor
     */
    fullName?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOProfesor
     */
    classCount?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOProfesor
     */
    enabledStatus?: string;
}


/**
 * @export
 */
export const DTOProfesorRoleEnum = {
    Admin: 'ADMIN',
    Profesor: 'PROFESOR',
    Alumno: 'ALUMNO',
    Usuario: 'USUARIO'
} as const;
export type DTOProfesorRoleEnum = typeof DTOProfesorRoleEnum[keyof typeof DTOProfesorRoleEnum];